+++ kamehouse-groot/src/main/webapp/kame-house-groot/js/admin/kamehouse-shell/kamehouse-shell-index.js
@@ -123,6 +123,7 @@ class KameHouseShellLoader {
       attr: {
         class: "kamehouse-shell-table-btn",
       },
+      mobileClass: null,
       backgroundImg: null,
       html: scriptName,
       data: {