+++ kamehouse-ui/src/main/webapp/kamehouse/js/plugin/kamehouse-crud-manager.js
@@ -839,6 +839,7 @@ class CrudManager {
       attr: {
         class: "img-btn-kh m-15-d-r-kh",
       },
+      mobileClass: null,
       backgroundImg: "/kame-house/img/other/edit.png",
       html: null,
       data: {
@@ -859,6 +860,7 @@ class CrudManager {
       attr: {
         class: "img-btn-kh",
       },
+      mobileClass: null,
       backgroundImg: "/kame-house/img/other/delete.png",
       html: null,
       data: {
@@ -878,6 +880,7 @@ class CrudManager {
       attr: {
         class: "form-submit-btn-kh",
       },
+      mobileClass: null,
       backgroundImg: null,
       html: "Yes",
       data: {
@@ -1116,6 +1119,7 @@ class CrudManager {
         id: addButtonId,
         class: "img-btn-kh p-7-d-kh m-7-d-kh",
       },
+      mobileClass: null,
       backgroundImg: "/kame-house/img/other/add-gray-dark.png",
       html: null,
       data: null,
@@ -1128,6 +1132,7 @@ class CrudManager {
         id: removeButtonId,
         class: "img-btn-kh p-7-d-kh m-7-d-kh",
       },
+      mobileClass: null,
       backgroundImg: "/kame-house/img/other/remove-gray-dark.png",
       html: null,
       data: null,